operate/agentic_chaos/requirements.txt
google-adk>=1.27
aiosqlite>=0.20
pydantic>=2.0
